Proper refactor of TurnTakingCollapsed
