+++ client/src/App.js
@@ -17,7 +17,6 @@ import Dashboard from "./components/dashboard/Dashboard";
 import TalkRatio from './components/visualizations/TalkRatio';
 import Transcript from './components/visualizations/Transcript';
 import TurnTaking from './components/visualizations/turn-taking/TurnTaking';
-import TurnTakingCollapsed from './components/visualizations/turn-taking/TurnTakingCollapsed';
 
 // Check for token to keep user logged in
 if (localStorage.jwtToken) {
@@ -97,7 +96,7 @@ class App extends Component {
             <Switch>
               <PrivateRoute exact path="/dashboard" component={Dashboard} />
               <PrivateRoute exact path="/dashboard/talk-ratio" component={TalkRatio} />
-              <PrivateRoute exact path="/dashboard/turn-taking" component={TurnTakingCollapsed} />
+              <PrivateRoute exact path="/dashboard/turn-taking" component={TurnTaking} />
               <PrivateRoute exact path="/dashboard/transcript" component={Transcript} />
             </Switch>
           </div>
