+++ client/src/components/visualizations/turn-taking/TurnTakingCollapsed.js
@@ -1,21 +1,60 @@
 import React, { Component } from 'react';
-import CollapsedBar from './CollapsedBar';
-
 import Parser from '../../../data/parser';
+import LegendLabels from '../../../fixtures/legend_labels';
 
 export default class TurnTakingCollapsed extends Component {
+    legendLabels = LegendLabels;
+
     render() {
         var chartData = Parser.parsedData();
 
         return (
-            <div className="turn-taking-visualization-container">
-              <div className="turn-taking-visualization">
-                {chartData.map((item, index) => {
-                    return (
-                      <CollapsedBar key={index} data={item} />
-                    )
-                })}
-              </div>
+            <div className="turn-taking-bars-collapsed-visualization">
+              {chartData.map((item, index) => {
+                  var isStudentData = item.speaker.includes("Student"),
+                      isTeacherData = item.speaker === "Teacher";
+
+                  var legendLabelValue = item.types[item.types.length - 1];
+                  var barColor = this.legendLabels.find(item => item.value === legendLabelValue).color;
+                  var barBorder = "";
+                  var boxSizing = "";
+                  if (item.types.length > 1) { // if it has multiple types, draw a border around the bar
+                      var borderValue = item.types && item.types[0];
+                      barBorder = `1px solid ${this.legendLabels.find(item => item.value === borderValue)}.color`;
+                      boxSizing = "border-box";
+                  }
+
+                  var barWidth = item.length,
+                      barHeight = "3px";
+
+                  var baseStyle = { height: barHeight },
+                      extendedStyle = { backgroundColor: barColor, border: barBorder, boxSizing: boxSizing, width: barWidth },
+                      teacherStyle = {},
+                      studentStyle = {};
+
+                  if (isTeacherData) {
+                      studentStyle = baseStyle;
+                      teacherStyle = { ...baseStyle, ...extendedStyle };
+                  }
+
+                  if (isStudentData) {
+                      studentStyle = { ...baseStyle, ...extendedStyle };
+                      teacherStyle = baseStyle;
+                  }
+
+                  return (
+                      <div className="turn-taking-bar-collapsed">
+                        <div className="turn-taking-bar-teacher-outer">
+                          <div className="turn-taking-bar-teacher-inner" style={teacherStyle}>
+                          </div>
+                        </div>
+                        <div className="turn-taking-bar-student-outer">
+                          <div className="turn-taking-bar-student-inner" style={studentStyle}>
+                          </div>
+                        </div>
+                      </div>
+                  );
+              })}
             </div>
         );
     }
