+++ client/src/styles/index.scss
@@ -3,7 +3,7 @@
 @import "./forms.scss";
 @import "./legend.scss";
 @import "./navbar.scss";
-// @import "./turn_taking.scss";
+@import "./turn_taking.scss";
 @import "./turn_taking_collapsed.scss";
 @import "./transcript.scss";
